requirements.txt
Django==5.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
Faker==21.0.0